+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 [Unreleased]: https://github.com/chaostoolkit-incubator/chaostoolkit-opentracing/compare/0.3.1...HEAD
 
+### Added
+
+-   Headers support for OLTP HTTP/GRPC exporters
+
 ## [0.3.1][] - 2021-12-13
 
 [0.3.1]: https://github.com/chaostoolkit-incubator/chaostoolkit-opentracing/compare/0.3.0...0.3.1